consolidate dapr renderers/handlers
