+++ pkg/radrp/outputresource/local_ids.go
@@ -45,13 +45,11 @@ const (
 	LocalIDAzureServiceBusQueue         = "AzureServiceBusQueue"
 	LocalIDAzureSqlServer               = "AzureSqlServer"
 	LocalIDAzureSqlServerDatabase       = "AzureSqlServerDatabase"
-	LocalIDExtender                      = "Extender"
-	LocalIDDaprPubSubGeneric            = "DaprPubSubGeneric"
+	LocalIDExtender                     = "Extender"
 	LocalIDDaprStateStoreAzureStorage   = "DaprStateStoreAzureStorage"
 	LocalIDDaprStateStoreComponent      = "DaprStateStoreComponent"
 	LocalIDDaprStateStoreSQLServer      = "DaprStateStoreSQLServer"
-	LocalIDDaprStateStoreGeneric        = "DaprStateStoreGeneric"
-	LocalIDDaprSecretStoreGeneric       = "DaprSecretStoreGeneric"
+	LocalIDDaprComponent                = "DaprComponent"
 	LocalIDDeployment                   = "Deployment"
 	LocalIDGateway                      = "Gateway"
 	LocalIDHttpRoute                    = "HttpRoute"
