+++ pkg/renderers/daprpubsubv1alpha3/azurerender.go
@@ -7,7 +7,6 @@ package daprpubsubv1alpha3
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 
@@ -18,7 +17,6 @@ import (
 	"github.com/project-radius/radius/pkg/providers"
 	"github.com/project-radius/radius/pkg/radrp/outputresource"
 	"github.com/project-radius/radius/pkg/renderers"
-	"github.com/project-radius/radius/pkg/renderers/dapr"
 	"github.com/project-radius/radius/pkg/resourcekinds"
 	"github.com/project-radius/radius/pkg/resourcemodel"
 )
@@ -30,7 +28,7 @@ type PubSubFunc = func(renderers.RendererResource) (renderers.RendererOutput, er
 // SupportedAzurePubSubKindValues is a map of supported resource kinds for Azure and the associated renderer
 var SupportedAzurePubSubKindValues = map[string]PubSubFunc{
 	resourcekinds.DaprPubSubTopicAzureServiceBus: GetDaprPubSubAzureServiceBus,
-	resourcekinds.DaprPubSubTopicGeneric:         GetDaprPubSubAzureGeneric,
+	resourcekinds.DaprGeneric:                    GetDaprPubSubGeneric,
 }
 
 type Renderer struct {
@@ -103,55 +101,6 @@ func GetDaprPubSubAzureServiceBus(resource renderers.RendererResource) (renderer
 		Resources:      []outputresource.OutputResource{output},
 		ComputedValues: values,
 		SecretValues:   secrets,
-	}, nil
-}
-
-func GetDaprPubSubAzureGeneric(resource renderers.RendererResource) (renderers.RendererOutput, error) {
-	properties := radclient.DaprPubSubTopicGenericResourceProperties{}
-	err := resource.ConvertDefinition(&properties)
-	if err != nil {
-		return renderers.RendererOutput{}, err
-	}
-
-	daprGeneric := dapr.DaprGeneric{
-		Type:     properties.Type,
-		Version:  properties.Version,
-		Metadata: properties.Metadata,
-	}
-
-	err = dapr.ValidateDaprGenericObject(daprGeneric)
-	if err != nil {
-		return renderers.RendererOutput{}, err
-	}
-
-	// Convert metadata to string
-	metadataSerialized, err := json.Marshal(properties.Metadata)
-	if err != nil {
-		return renderers.RendererOutput{}, err
-	}
-
-	output := outputresource.OutputResource{
-		LocalID: outputresource.LocalIDDaprPubSubGeneric,
-		ResourceType: resourcemodel.ResourceType{
-			Type:     resourcekinds.DaprPubSubTopicGeneric,
-			Provider: providers.ProviderKubernetes,
-		},
-		Resource: map[string]string{
-			handlers.ResourceName:            resource.ResourceName,
-			handlers.KubernetesNamespaceKey:  resource.ApplicationName,
-			handlers.KubernetesAPIVersionKey: "dapr.io/v1alpha1",
-			handlers.KubernetesKindKey:       "Component",
-
-			handlers.GenericDaprTypeKey:     *properties.Type,
-			handlers.GenericDaprVersionKey:  *properties.Version,
-			handlers.GenericDaprMetadataKey: string(metadataSerialized),
-		},
-	}
-
-	return renderers.RendererOutput{
-		Resources:      []outputresource.OutputResource{output},
-		ComputedValues: nil,
-		SecretValues:   nil,
 	}, nil
 }
