+++ pkg/renderers/daprpubsubv1alpha3/azurerender_test.go
@@ -7,7 +7,6 @@ package daprpubsubv1alpha3
 
 import (
 	"context"
-	"encoding/json"
 	"testing"
 
 	"github.com/project-radius/radius/pkg/handlers"
@@ -72,117 +71,4 @@ func Test_Render_InvalidResourceType(t *testing.T) {
 	_, err := renderer.Render(context.Background(), renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
 	require.Error(t, err)
 	require.Equal(t, "the 'resource' field must refer to a ServiceBus Topic", err.Error())
-}
-
-func Test_Render_Generic_Success(t *testing.T) {
-	renderer := Renderer{}
-
-	dependencies := map[string]renderers.RendererDependency{}
-	resource := renderers.RendererResource{
-		ApplicationName: "test-app",
-		ResourceName:    "test-resource",
-		ResourceType:    ResourceType,
-		Definition: map[string]interface{}{
-			"kind": resourcekinds.DaprPubSubTopicGeneric,
-			"type": "pubsub.kafka",
-			"metadata": map[string]interface{}{
-				"foo": "bar",
-			},
-			"version": "v1",
-		},
-	}
-
-	renderer.PubSubs = SupportedAzurePubSubKindValues
-	result, err := renderer.Render(context.Background(), renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
-	require.NoError(t, err)
-
-	require.Len(t, result.Resources, 1)
-	output := result.Resources[0]
-
-	require.Equal(t, outputresource.LocalIDDaprPubSubGeneric, output.LocalID)
-	require.Equal(t, resourcekinds.DaprPubSubTopicGeneric, output.ResourceType.Type)
-
-	metadata := map[string]interface{}{
-		"foo": "bar",
-	}
-	metadataSerialized, err := json.Marshal(metadata)
-	require.NoError(t, err, "Could not serialize metadata")
-
-	expected := map[string]string{
-		handlers.ResourceName:            resource.ResourceName,
-		handlers.KubernetesNamespaceKey:  resource.ApplicationName,
-		handlers.KubernetesAPIVersionKey: "dapr.io/v1alpha1",
-		handlers.KubernetesKindKey:       "Component",
-
-		handlers.GenericDaprTypeKey:     "pubsub.kafka",
-		handlers.GenericDaprVersionKey:  "v1",
-		handlers.GenericDaprMetadataKey: string(metadataSerialized),
-	}
-	require.Equal(t, expected, output.Resource)
-}
-
-func Test_Render_Generic_MissingMetadata(t *testing.T) {
-	renderer := Renderer{}
-
-	dependencies := map[string]renderers.RendererDependency{}
-	resource := renderers.RendererResource{
-		ApplicationName: "test-app",
-		ResourceName:    "test-resource",
-		ResourceType:    ResourceType,
-		Definition: map[string]interface{}{
-			"kind":     resourcekinds.DaprPubSubTopicGeneric,
-			"type":     "pubsub.kafka",
-			"metadata": map[string]string{},
-			"version":  "v1",
-		},
-	}
-
-	renderer.PubSubs = SupportedAzurePubSubKindValues
-	_, err := renderer.Render(context.Background(), renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
-	require.Error(t, err)
-	require.Equal(t, "No metadata specified for Dapr component of type pubsub.kafka", err.Error())
-}
-
-func Test_Render_Generic_MissingType(t *testing.T) {
-	renderer := Renderer{}
-
-	dependencies := map[string]renderers.RendererDependency{}
-	resource := renderers.RendererResource{
-		ApplicationName: "test-app",
-		ResourceName:    "test-resource",
-		ResourceType:    ResourceType,
-		Definition: map[string]interface{}{
-			"kind":     resourcekinds.DaprPubSubTopicGeneric,
-			"type":     "",
-			"metadata": map[string]string{},
-			"version":  "v1",
-		},
-	}
-
-	renderer.PubSubs = SupportedAzurePubSubKindValues
-	_, err := renderer.Render(context.Background(), renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
-	require.Error(t, err)
-	require.Equal(t, "No type specified for generic Dapr component", err.Error())
-}
-
-func Test_Render_Generic_MissingVersion(t *testing.T) {
-	renderer := Renderer{}
-
-	dependencies := map[string]renderers.RendererDependency{}
-	resource := renderers.RendererResource{
-		ApplicationName: "test-app",
-		ResourceName:    "test-resource",
-		ResourceType:    ResourceType,
-		Definition: map[string]interface{}{
-			"kind":     resourcekinds.DaprPubSubTopicGeneric,
-			"type":     "pubsub.kafka",
-			"metadata": map[string]string{},
-			"version":  "",
-		},
-	}
-
-	renderer.PubSubs = SupportedAzurePubSubKindValues
-	_, err := renderer.Render(context.Background(), renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
-	require.Error(t, err)
-	require.Equal(t, "No Dapr component version specified for generic Dapr component", err.Error())
 }
