+++ pkg/renderers/daprpubsubv1alpha3/kubernetesrender.go
@@ -11,20 +11,15 @@ import (
 	"fmt"
 
 	"github.com/project-radius/radius/pkg/azure/azresources"
-	"github.com/project-radius/radius/pkg/azure/radclient"
-	"github.com/project-radius/radius/pkg/providers"
-	"github.com/project-radius/radius/pkg/radrp/outputresource"
 	"github.com/project-radius/radius/pkg/renderers"
-	"github.com/project-radius/radius/pkg/renderers/dapr"
 	"github.com/project-radius/radius/pkg/resourcekinds"
-	"github.com/project-radius/radius/pkg/resourcemodel"
 )
 
 var _ renderers.Renderer = (*KubernetesRenderer)(nil)
 
 // SupportedKubernetesPubSubKindValues is a map of supported resource kinds for k8s and the associated renderer
 var SupportedKubernetesPubSubKindValues = map[string]PubSubFunc{
-	resourcekinds.DaprPubSubTopicGeneric: GetDaprPubSubAzureGenericKubernetes,
+	resourcekinds.DaprGeneric: GetDaprPubSubGeneric,
 }
 
 type KubernetesRenderer struct {
@@ -33,45 +28,6 @@ type KubernetesRenderer struct {
 
 func (r *KubernetesRenderer) GetDependencyIDs(ctx context.Context, workload renderers.RendererResource) ([]azresources.ResourceID, []azresources.ResourceID, error) {
 	return nil, nil, nil
-}
-
-func GetDaprPubSubAzureGenericKubernetes(resource renderers.RendererResource) (renderers.RendererOutput, error) {
-	properties := radclient.DaprPubSubTopicGenericResourceProperties{}
-	err := resource.ConvertDefinition(&properties)
-	if err != nil {
-		return renderers.RendererOutput{}, err
-	}
-
-	daprGeneric := dapr.DaprGeneric{
-		Type:     properties.Type,
-		Version:  properties.Version,
-		Metadata: properties.Metadata,
-	}
-
-	err = dapr.ValidateDaprGenericObject(daprGeneric)
-	if err != nil {
-		return renderers.RendererOutput{}, err
-	}
-
-	pubsubResource, err := dapr.ConstructDaprGeneric(daprGeneric, resource.ApplicationName, resource.ResourceName)
-	if err != nil {
-		return renderers.RendererOutput{}, err
-	}
-
-	output := outputresource.OutputResource{
-		LocalID: outputresource.LocalIDDaprPubSubGeneric,
-		ResourceType: resourcemodel.ResourceType{
-			Type:     resourcekinds.DaprPubSubGeneric,
-			Provider: providers.ProviderKubernetes,
-		},
-		Resource: &pubsubResource,
-	}
-
-	return renderers.RendererOutput{
-		Resources:      []outputresource.OutputResource{output},
-		ComputedValues: nil,
-		SecretValues:   nil,
-	}, nil
 }
 
 func (r *KubernetesRenderer) Render(ctx context.Context, options renderers.RenderOptions) (renderers.RendererOutput, error) {
