+++ pkg/renderers/daprpubsubv1alpha3/render_test.go
@@ -1,8 +1,3 @@
-// ------------------------------------------------------------
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-// ------------------------------------------------------------
-
 package daprpubsubv1alpha3
 
 import (
@@ -22,30 +17,16 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
-const (
-	appName           = "test-app"
-	resourceName      = "test-resource"
-	pubsubType        = "pubsub.kafka"
-	daprPubSubVersion = "v1"
-	daprVersion       = "dapr.io/v1alpha1"
-	k8sKind           = "Component"
-)
-
-func initRenderer() Renderer {
-	return Renderer{
-		PubSubs: SupportedKubernetesPubSubKindValues,
-	}
-}
-
-func Test_Render_Generic_Kubernetes_Success(t *testing.T) {
+func Test_Render_Generic_Success(t *testing.T) {
+	renderer := Renderer{}
 
 	dependencies := map[string]renderers.RendererDependency{}
 	resource := renderers.RendererResource{
 		ApplicationName: "test-app",
 		ResourceName:    "test-resource",
-		ResourceType:    resourcekinds.DaprPubSubGeneric,
+		ResourceType:    resourcekinds.DaprGeneric,
 		Definition: map[string]interface{}{
-			"kind": resourcekinds.DaprPubSubTopicGeneric,
+			"kind": resourcekinds.DaprGeneric,
 			"type": "pubsub.kafka",
 			"metadata": map[string]interface{}{
 				"foo": "bar",
@@ -54,17 +35,17 @@ func Test_Render_Generic_Kubernetes_Success(t *testing.T) {
 		},
 	}
 
-	renderer := initRenderer()
+	renderer.PubSubs = SupportedAzurePubSubKindValues
 	result, err := renderer.Render(context.Background(), renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
 	require.NoError(t, err)
 
 	require.Len(t, result.Resources, 1)
 	output := result.Resources[0]
 
-	require.Equal(t, outputresource.LocalIDDaprPubSubGeneric, output.LocalID)
-	require.Equal(t, resourcekinds.DaprPubSubGeneric, output.ResourceType.Type)
+	require.Equal(t, outputresource.LocalIDDaprComponent, output.LocalID)
+	require.Equal(t, resourcekinds.DaprComponent, output.ResourceType.Type)
 
-	expected := unstructured.Unstructured{
+	expected := &unstructured.Unstructured{
 		Object: map[string]interface{}{
 			"apiVersion": daprVersion,
 			"kind":       k8sKind,
@@ -85,64 +66,70 @@ func Test_Render_Generic_Kubernetes_Success(t *testing.T) {
 			},
 		},
 	}
-	require.Equal(t, &expected, output.Resource)
+	require.Equal(t, expected, output.Resource)
 }
 
-func Test_Render_Generic_Kubernetes_MissingMetadata(t *testing.T) {
+func Test_Render_Generic_MissingMetadata(t *testing.T) {
+	renderer := Renderer{}
+
 	dependencies := map[string]renderers.RendererDependency{}
 	resource := renderers.RendererResource{
 		ApplicationName: "test-app",
 		ResourceName:    "test-resource",
 		ResourceType:    ResourceType,
 		Definition: map[string]interface{}{
-			"kind":     resourcekinds.DaprPubSubTopicGeneric,
+			"kind":     resourcekinds.DaprGeneric,
 			"type":     "pubsub.kafka",
 			"metadata": map[string]string{},
 			"version":  "v1",
 		},
 	}
 
-	renderer := initRenderer()
+	renderer.PubSubs = SupportedAzurePubSubKindValues
 	_, err := renderer.Render(context.Background(), renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
 	require.Error(t, err)
 	require.Equal(t, "No metadata specified for Dapr component of type pubsub.kafka", err.Error())
 }
 
-func Test_Render_Generic_Kubernetes_MissingType(t *testing.T) {
+func Test_Render_Generic_MissingType(t *testing.T) {
+	renderer := Renderer{}
+
 	dependencies := map[string]renderers.RendererDependency{}
 	resource := renderers.RendererResource{
 		ApplicationName: "test-app",
 		ResourceName:    "test-resource",
 		ResourceType:    ResourceType,
 		Definition: map[string]interface{}{
-			"kind":     resourcekinds.DaprPubSubTopicGeneric,
+			"kind":     resourcekinds.DaprGeneric,
 			"type":     "",
 			"metadata": map[string]string{},
 			"version":  "v1",
 		},
 	}
 
-	renderer := initRenderer()
+	renderer.PubSubs = SupportedAzurePubSubKindValues
 	_, err := renderer.Render(context.Background(), renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
 	require.Error(t, err)
 	require.Equal(t, "No type specified for generic Dapr component", err.Error())
 }
 
-func Test_Render_Generic_Kubernetes_MissingVersion(t *testing.T) {
+func Test_Render_Generic_MissingVersion(t *testing.T) {
+	renderer := Renderer{}
+
 	dependencies := map[string]renderers.RendererDependency{}
 	resource := renderers.RendererResource{
 		ApplicationName: "test-app",
 		ResourceName:    "test-resource",
 		ResourceType:    ResourceType,
 		Definition: map[string]interface{}{
-			"kind":     resourcekinds.DaprPubSubTopicGeneric,
+			"kind":     resourcekinds.DaprGeneric,
 			"type":     "pubsub.kafka",
 			"metadata": map[string]string{},
 			"version":  "",
 		},
 	}
 
-	renderer := initRenderer()
+	renderer.PubSubs = SupportedAzurePubSubKindValues
 	_, err := renderer.Render(context.Background(), renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
 	require.Error(t, err)
 	require.Equal(t, "No Dapr component version specified for generic Dapr component", err.Error())
