+++ pkg/renderers/daprsecretstorev1alpha3/render_test.go
@@ -7,12 +7,10 @@ package daprsecretstorev1alpha3
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"testing"
 
 	"github.com/go-logr/logr"
-	"github.com/project-radius/radius/pkg/handlers"
 	"github.com/project-radius/radius/pkg/kubernetes"
 	"github.com/project-radius/radius/pkg/radlogger"
 	"github.com/project-radius/radius/pkg/radrp/outputresource"
@@ -39,53 +37,6 @@ func createContext(t *testing.T) context.Context {
 	return logr.NewContext(context.Background(), logger)
 }
 
-func Test_Render_Azure_Generic_Success(t *testing.T) {
-	ctx := createContext(t)
-	renderer := Renderer{SupportedAzureSecretStoreKindValues}
-
-	dependencies := map[string]renderers.RendererDependency{}
-	resource := renderers.RendererResource{
-		ApplicationName: "test-app",
-		ResourceName:    "test-resource",
-		ResourceType:    ResourceType,
-		Definition: map[string]interface{}{
-			"kind":    "generic",
-			"type":    "secretstores.azure.keyvault",
-			"version": "v1",
-			"metadata": map[string]interface{}{
-				"vaultName": "testVault",
-			},
-		},
-	}
-
-	result, err := renderer.Render(ctx, renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
-	require.NoError(t, err)
-
-	require.Len(t, result.Resources, 1)
-	output := result.Resources[0]
-
-	require.Equal(t, outputresource.LocalIDDaprSecretStoreGeneric, output.LocalID)
-	require.Equal(t, resourcekinds.DaprSecretStoreGeneric, output.ResourceType.Type)
-
-	metadata := map[string]interface{}{
-		"vaultName": "testVault",
-	}
-	metadataSerialized, err := json.Marshal(metadata)
-	require.NoError(t, err, "Could not serialize metadata")
-
-	expected := map[string]string{
-		handlers.KubernetesNameKey:       "test-resource",
-		handlers.KubernetesNamespaceKey:  "test-app",
-		handlers.KubernetesAPIVersionKey: "dapr.io/v1alpha1",
-		handlers.KubernetesKindKey:       "Component",
-		handlers.ResourceName:            "test-resource",
-		handlers.GenericDaprTypeKey:      "secretstores.azure.keyvault",
-		handlers.GenericDaprVersionKey:   "v1",
-		handlers.GenericDaprMetadataKey:  string(metadataSerialized),
-	}
-	require.Equal(t, expected, output.Resource)
-}
-
 func Test_Render_UnsupportedKind(t *testing.T) {
 	ctx := createContext(t)
 	renderer := Renderer{SupportedAzureSecretStoreKindValues}
@@ -105,75 +56,7 @@ func Test_Render_UnsupportedKind(t *testing.T) {
 	require.Equal(t, fmt.Sprintf("azure.keyvault is not supported. Supported kind values: %s", getAlphabeticallySortedKeys(SupportedAzureSecretStoreKindValues)), err.Error())
 }
 
-func Test_Render_Azure_Generic_MissingMetadata(t *testing.T) {
-	ctx := createContext(t)
-	renderer := Renderer{SupportedAzureSecretStoreKindValues}
-
-	dependencies := map[string]renderers.RendererDependency{}
-	resource := renderers.RendererResource{
-		ApplicationName: "test-app",
-		ResourceName:    "test-resource",
-		ResourceType:    ResourceType,
-		Definition: map[string]interface{}{
-			"kind":     "generic",
-			"type":     "azure.keyvault",
-			"version":  "v1",
-			"metadata": map[string]interface{}{},
-		},
-	}
-
-	_, err := renderer.Render(ctx, renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
-	require.Error(t, err)
-	require.Equal(t, "No metadata specified for Dapr component of type azure.keyvault", err.Error())
-}
-
-func Test_Render_Azure_Generic_MissingType(t *testing.T) {
-	ctx := createContext(t)
-	renderer := Renderer{SupportedAzureSecretStoreKindValues}
-
-	dependencies := map[string]renderers.RendererDependency{}
-	resource := renderers.RendererResource{
-		ApplicationName: "test-app",
-		ResourceName:    "test-resource",
-		ResourceType:    ResourceType,
-		Definition: map[string]interface{}{
-			"kind":    "generic",
-			"version": "v1",
-			"metadata": map[string]interface{}{
-				"foo": "bar",
-			},
-		},
-	}
-
-	_, err := renderer.Render(ctx, renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
-	require.Error(t, err)
-	require.Equal(t, "No type specified for generic Dapr component", err.Error())
-}
-
-func Test_Render_Azure_Generic_MissingVersion(t *testing.T) {
-	ctx := createContext(t)
-	renderer := Renderer{SupportedAzureSecretStoreKindValues}
-
-	dependencies := map[string]renderers.RendererDependency{}
-	resource := renderers.RendererResource{
-		ApplicationName: "test-app",
-		ResourceName:    "test-resource",
-		ResourceType:    ResourceType,
-		Definition: map[string]interface{}{
-			"kind": "generic",
-			"type": "azure.keyvault",
-			"metadata": map[string]interface{}{
-				"foo": "bar",
-			},
-		},
-	}
-
-	_, err := renderer.Render(ctx, renderers.RenderOptions{Resource: resource, Dependencies: dependencies})
-	require.Error(t, err)
-	require.Equal(t, "No Dapr component version specified for generic Dapr component", err.Error())
-}
-
-func Test_Render_Kubernetes_Generic_Success(t *testing.T) {
+func Test_Render_Generic_Success(t *testing.T) {
 	ctx := createContext(t)
 	renderer := Renderer{SupportedKubernetesSecretStoreKindValues}
 
@@ -183,7 +66,7 @@ func Test_Render_Kubernetes_Generic_Success(t *testing.T) {
 		ResourceName:    resourceName,
 		ResourceType:    ResourceType,
 		Definition: map[string]interface{}{
-			"kind":    "generic",
+			"kind":    resourcekinds.DaprGeneric,
 			"type":    "secretstores.kubernetes",
 			"version": "v1",
 			"metadata": map[string]interface{}{
@@ -198,8 +81,8 @@ func Test_Render_Kubernetes_Generic_Success(t *testing.T) {
 	require.Len(t, result.Resources, 1)
 	output := result.Resources[0]
 
-	require.Equal(t, outputresource.LocalIDDaprSecretStoreGeneric, output.LocalID)
-	require.Equal(t, resourcekinds.DaprSecretStoreGeneric, output.ResourceType.Type)
+	require.Equal(t, outputresource.LocalIDDaprComponent, output.LocalID)
+	require.Equal(t, resourcekinds.DaprComponent, output.ResourceType.Type)
 
 	expected := unstructured.Unstructured{
 		Object: map[string]interface{}{
@@ -225,7 +108,7 @@ func Test_Render_Kubernetes_Generic_Success(t *testing.T) {
 	require.Equal(t, &expected, output.Resource)
 }
 
-func Test_Render_Kubernetes_Generic_MissingMetadata(t *testing.T) {
+func Test_Render_Generic_MissingMetadata(t *testing.T) {
 	ctx := createContext(t)
 	renderer := Renderer{SupportedKubernetesSecretStoreKindValues}
 
@@ -235,7 +118,7 @@ func Test_Render_Kubernetes_Generic_MissingMetadata(t *testing.T) {
 		ResourceName:    resourceName,
 		ResourceType:    ResourceType,
 		Definition: map[string]interface{}{
-			"kind":     "generic",
+			"kind":     resourcekinds.DaprGeneric,
 			"type":     "secretstores.kubernetes",
 			"version":  daprSecretStoreVersion,
 			"metadata": map[string]interface{}{},
@@ -247,7 +130,7 @@ func Test_Render_Kubernetes_Generic_MissingMetadata(t *testing.T) {
 	require.Equal(t, "No metadata specified for Dapr component of type secretstores.kubernetes", err.Error())
 }
 
-func Test_Render_Kubernetes_Generic_MissingType(t *testing.T) {
+func Test_Render_Generic_MissingType(t *testing.T) {
 	ctx := createContext(t)
 	renderer := Renderer{SupportedKubernetesSecretStoreKindValues}
 
@@ -257,7 +140,7 @@ func Test_Render_Kubernetes_Generic_MissingType(t *testing.T) {
 		ResourceName:    resourceName,
 		ResourceType:    ResourceType,
 		Definition: map[string]interface{}{
-			"kind":    "generic",
+			"kind":    resourcekinds.DaprGeneric,
 			"version": "v1",
 			"metadata": map[string]interface{}{
 				"foo": "bar",
@@ -270,7 +153,7 @@ func Test_Render_Kubernetes_Generic_MissingType(t *testing.T) {
 	require.Equal(t, "No type specified for generic Dapr component", err.Error())
 }
 
-func Test_Render_Kubernetes_Generic_MissingVersion(t *testing.T) {
+func Test_Render_Generic_MissingVersion(t *testing.T) {
 	ctx := createContext(t)
 	renderer := Renderer{SupportedKubernetesSecretStoreKindValues}
 
@@ -280,7 +163,7 @@ func Test_Render_Kubernetes_Generic_MissingVersion(t *testing.T) {
 		ResourceName:    resourceName,
 		ResourceType:    ResourceType,
 		Definition: map[string]interface{}{
-			"kind": "generic",
+			"kind": resourcekinds.DaprGeneric,
 			"type": "secretstores.kubernetes",
 			"metadata": map[string]interface{}{
 				"foo": "bar",
