+++ pkg/resourcekinds/resource_kinds.go
@@ -19,11 +19,9 @@ const (
 	StatefulSet                      = "StatefulSet"
 	SecretProviderClass              = "SecretProviderClass"
 	DaprStateStoreAzureStorage       = "dapr.statestore.azurestorage"
-	DaprPubSubTopicGeneric           = "generic"
+	DaprGeneric                      = "generic"
+	DaprComponent                    = "dapr.io.component"
 	DaprPubSubTopicAzureServiceBus   = "pubsub.azure.servicebus"
-	DaprStateStoreGeneric            = "dapr.io.statestore"
-	DaprSecretStoreGeneric           = "dapr.io.secretstore"
-	DaprPubSubGeneric                = "dapr.io.pubsubtopic"
 	AzureCosmosAccount               = "azure.cosmosdb.account"
 	AzureCosmosDBMongo               = "azure.cosmosdb.mongo"
 	AzureCosmosDBSQL                 = "azure.cosmosdb.sql"
