+++ test/functional/azure/resources/dapr_secretstore_generic_test.go
@@ -42,7 +42,7 @@ func Test_DaprSecretStoreGeneric(t *testing.T) {
 						ResourceName:    "secretstore-generic",
 						ResourceType:    daprsecretstorev1alpha3.ResourceType,
 						OutputResources: map[string]validation.ExpectedOutputResource{
-							outputresource.LocalIDDaprSecretStoreGeneric: validation.NewOutputResource(outputresource.LocalIDDaprSecretStoreGeneric, rest.ResourceType{Type: resourcekinds.DaprSecretStoreGeneric, Provider: providers.ProviderKubernetes}, false, rest.OutputResourceStatus{}),
+							outputresource.LocalIDDaprComponent: validation.NewOutputResource(outputresource.LocalIDDaprComponent, rest.ResourceType{Type: resourcekinds.DaprComponent, Provider: providers.ProviderKubernetes}, false, rest.OutputResourceStatus{}),
 						},
 					},
 				},
