+++ test/functional/kubernetes/resources/dapr_statestore_test.go
@@ -35,7 +35,7 @@ func Test_DaprStateStore_Generic(t *testing.T) {
 						ApplicationName: application,
 						ResourceName:    "statestore",
 						OutputResources: map[string]validation.ExpectedOutputResource{
-							outputresource.LocalIDDaprStateStoreGeneric: validation.NewOutputResource(outputresource.LocalIDDaprStateStoreGeneric, rest.ResourceType{Type: resourcekinds.DaprStateStoreGeneric, Provider: providers.ProviderKubernetes}, false, rest.OutputResourceStatus{}),
+							outputresource.LocalIDDaprComponent: validation.NewOutputResource(outputresource.LocalIDDaprComponent, rest.ResourceType{Type: resourcekinds.DaprComponent, Provider: providers.ProviderKubernetes}, false, rest.OutputResourceStatus{}),
 						},
 					},
 				},
